lint: move metric fns to package
